BSS-104
Ethereum value parser utility
